% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessQualityOfcaptWKL.R
\name{assessQualityOfcaptWKL}
\alias{assessQualityOfcaptWKL}
\title{Assess how well model captured PWLs}
\usage{
assessQualityOfcaptWKL(VData, band, stabilityindex = "p_unstable")
}
\arguments{
\item{VData}{object; don't include both primary/secondary and tertiary gtype_ranks!}
}
\description{
This function calculates several indicators that help to assess how well the model captured captWKLs beyond simple layer structure.
It get's called by the overarching evaluation script \code{evaluate_main.R}
}
\details{
So, for each WKL the function computes
\itemize{
\item correlation factors for likelihood/distribution/sensitvity versus proportion of unstable grid points
\item the maximum proportion of unstable grid points during the lifetime of the WKL
\item the temporal lag between first forecaster concern (--> avy problem) and first time that grid points are unstable in that layer
\itemize{
\item different timing instances possible: first time (i) at least one grid point unstable, (ii) the majority of grid points is unstable, (iii) more than 50 \% of max proportion unstable
}
\item the temporal lag between latest forecaster concern (--> last day of avy problem) and last time that grid points are unstable in that layer
\itemize{
\item need to find rules for becoming dormant and waking up again versus becoming inactive
}
}
}
