% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignLayer2Datetag.R
\name{assignLayer2Datetag}
\alias{assignLayer2Datetag}
\title{Assign a layer to most representative datetag}
\usage{
assignLayer2Datetag(layer, potentialDatetags, silent = FALSE)
}
\arguments{
\item{layer}{snowprofileLayers-like object of one or multiple potential weak layers (with their ddate and bdate)}

\item{potentialDatetags}{array of Dates (as obtained by \link{derivePotentialDatetags})}

\item{silent}{if set to FALSE, the function will warn you if datetags seem extremely inappropriate for the layers' bdates}
}
\value{
This function will fill \code{layer$datetag} with the appropriate datetag
}
\description{
If you have a vector of potential date tags as obtained by \link{derivePotentialDatetags}, this function will
assign simulated layers to the optimal date tag based on following rules for layer assignment:
\itemize{
\item from all datetags that are older than the layer's ddate
\item the datetag that is closest to the layer's bdate will be assigned
}

If the layer formed after the oldest datetag that was provided, only rule 2 applies.

Note that if the layer is not buried yet (surface layer), the function will use the vdate (validation date, essentially the date of the profile) instead.
The vdate only exists for layers in the wklvalidation database, though, and the profile date cannot be queried by this function. Therefore,
if the function is used outside of the package context, assign \code{layer$vdate} as the profile date before calling this function!
}
