% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapConfusionMatrix.R
\name{bootstrapConfusionMatrix}
\alias{bootstrapConfusionMatrix}
\title{Compute confidence bands by bootstrapping}
\usage{
bootstrapConfusionMatrix(
  cm,
  perf = NULL,
  nboots = 4001,
  target = c("attributes", "scores")[1]
)
}
\arguments{
\item{cm}{confusion matrix data frame}

\item{perf}{performance object (obtained from \code{ROCR::performance}, only required for ROC and precision-recall confidence bands!)}

\item{nboots}{number of bootstrap replications}

\item{target}{compute confidence bands for performance curves or skill scores?}
}
\value{
data frame with confidence bands
}
\description{
Compute confidence bands by bootstrapping
}
