% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapConfusionMatrix.R
\name{calculateF1Score}
\alias{calculateF1Score}
\title{Compute F1 skill score...
...for a given cutoff-threshold.}
\usage{
calculateF1Score(sim, obs, cutoff)
}
\arguments{
\item{sim}{vector of numeric values between \link{0, 1} (e.g., proportion of unstable grid points)}

\item{obs}{vector of logicals (TRUE/FALSE) stating whether the layer was observed (of concern) or not.}

\item{cutoff}{in percentage within (0, 1]}
}
\value{
numeric value of skill score
}
\description{
Compute F1 skill score...
...for a given cutoff-threshold.
}
