% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateProportionUnstable.R
\name{calculateProportionCaptured}
\alias{calculateProportionCaptured}
\title{Calculate proportion of structurally captured grid points in ValidationDataSet VD
Same as \link{calculateProportionUnstable} but for structurally captured layers only. Hence, this function
is only applied to the vframe data.frame!}
\usage{
calculateProportionCaptured(VD)
}
\arguments{
\item{VD}{ValidationDataSet VD as obtained from \link{getVdata} from data base}
}
\description{
Calculate proportion of structurally captured grid points in ValidationDataSet VD
Same as \link{calculateProportionUnstable} but for structurally captured layers only. Hence, this function
is only applied to the vframe data.frame!
}
