% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateProportionUnstable.R
\name{calculateProportionUnstable}
\alias{calculateProportionUnstable}
\title{Calculate proportion of unstable grid points in ValidationDataSet VD}
\usage{
calculateProportionUnstable(VD, target = c("vframe", "poorWKL")[1])
}
\arguments{
\item{VD}{ValidationDataSet VD as obtained from \link{getVdata} from data base}

\item{target}{compute proportion for vframe, poorWKL or both?}
}
\description{
This function can compute the proportion for the vframe data.frame, or for the poorWKL data.frame.
In the vframe case, it identifies all captured and unstable layers that belong to each vframe row
and then computes their proportion relative to all grid points in the same subset. In the poorWKL
case, the function identifies all unstable layers that were not of concern, and computes their
proportion for each day and each model-derived date tag relative to all grid points at that day.
}
