% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeGsizeAtBurial.R
\name{computeGsizeAtBurial}
\alias{computeGsizeAtBurial}
\title{Compute median simulated grain size of layer at burial
This function can get called from the large evaluation script \code{evaluate_main.R} (by setting some config flags in the beginning of the script)
For each group of layers associated with a date tag it traces back the layers before their burial and then computes their median grain sizes.
Extremely CPU intense!!}
\usage{
computeGsizeAtBurial(VD)
}
\arguments{
\item{VD}{ValidationDataSet VD as obtained from \link{getVdata} from data base}
}
\value{
same VD object with variable \code{gsizeAtBurial_median} added to VD$vframe.
}
\description{
Compute median simulated grain size of layer at burial
This function can get called from the large evaluation script \code{evaluate_main.R} (by setting some config flags in the beginning of the script)
For each group of layers associated with a date tag it traces back the layers before their burial and then computes their median grain sizes.
Extremely CPU intense!!
}
