% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeLengthOfPreceedingDrySpell.R
\name{computeLengthOfPreceedingDrySpell}
\alias{computeLengthOfPreceedingDrySpell}
\title{Compute median simulated length of dry period before burial of a layer
This function can get called from the large evaluation script \code{evaluate_main.R} (by setting some config flags in the beginning of the script)
For each layer of concern, this function assembles the median length of the dry period before the burial of the layer.
To do so, it uses the information about the length of the search window that got computed for each layer of concern at each grid point
(during \code{validate_main}) and stored in VD$captWKL$tliml.}
\usage{
computeLengthOfPreceedingDrySpell(VD)
}
\arguments{
\item{VD}{ValidationDataSet VD as obtained from \link{getVdata} from data base}
}
\value{
same VD object with variable \code{drySpell_median} and \verb{*_std} added to VD$vframe.
}
\description{
Compute median simulated length of dry period before burial of a layer
This function can get called from the large evaluation script \code{evaluate_main.R} (by setting some config flags in the beginning of the script)
For each layer of concern, this function assembles the median length of the dry period before the burial of the layer.
To do so, it uses the information about the length of the search window that got computed for each layer of concern at each grid point
(during \code{validate_main}) and stored in VD$captWKL$tliml.
}
