% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapConfusionMatrix.R
\name{computePerformanceAttributesFromCM}
\alias{computePerformanceAttributesFromCM}
\title{Compute performance curves for bootstrap data subsets
i.e., helper function for bootstrapConfusionMatrix}
\usage{
computePerformanceAttributesFromCM(data, i)
}
\description{
Compute performance curves for bootstrap data subsets
i.e., helper function for bootstrapConfusionMatrix
}
