% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapConfusionMatrix.R
\name{computePerformanceScoresFromCM}
\alias{computePerformanceScoresFromCM}
\title{Compute skill scores for bootstrap data subsets
i.e., helper function for bootstrapConfusionMatrix}
\usage{
computePerformanceScoresFromCM(data, i = seq(data))
}
\description{
Compute skill scores for bootstrap data subsets
i.e., helper function for bootstrapConfusionMatrix
}
