% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBinteractions.R
\name{deleteExecutionFromDB}
\alias{deleteExecutionFromDB}
\title{Delete all data related to exec_uuid from entire database}
\usage{
deleteExecutionFromDB(
  exec_uuid,
  DBCon = NULL,
  DB = "Steed_DB",
  rf = FALSE,
  Renviron_path = ifelse(Sys.getenv("USER") == "flo",
    "/home/flo/documents/sfu/code/cedarstuff/CedarTunnel.Renviron",
    "/home/fherla/Cedar.Renviron"),
  verPlot_path = ifelse(DB == "Steed_DB",
    "/home/flo/documents/sfu/code/rpackages_SARP/sarp.2021.herla.snowprofilevalidation/output/figures/validationVerificationPlots/",
    "/home/fherla/project/fherla/snowprofilevalidation/output/figures/validationVerificationPlots/")
)
}
\description{
User will be prompted before final deletion!
}
\details{
remote usage prerequisite: on a terminal keep running SSH tunnel:
\verb{ssh -L 8888:cedar-mysql-vm.int.cedar.computecanada.ca:3306 fherla@cedar.computecanada.ca}
}
