% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivePotentialDatetags.R
\name{derivePotentialDatetags}
\alias{derivePotentialDatetags}
\title{Derive all potential date tags of a season from modeled weather data}
\usage{
derivePotentialDatetags(WX24, humanDatetags = NA)
}
\arguments{
\item{WX24}{data frame containing hn24, hn72, and rain24, rain72 for all days of the season and aggregated over all grid points.
hn24 and rain24 need to be provided as 75th percentiles, and hn72 and rain72 as 25th percentiles.}

\item{humanDatetags}{character or Date vector with date tags of layers of human concern}
}
\value{
an array of potential datetags as Dates
}
\description{
These model-derived date tags represent regional markers for times when potential critical avalanche layers got buried.
Analogously to human date tags, these regional markers are primarily represented by the onset dates of pronounced storms
with substantial snowfall, but these markers can also be dates when little but consistent snowfall amounts buried the snow surface.
}
\details{
To establish the model-derived date tags, this function first searches for dry periods by identifying all days with less than a trace amount
of snowfall or rain in the provided weather data. The actual model-derived date tags are defined by the start of substantial storm
periods following dry periods. However, since potentially critical layers can also be buried by small amounts of snow over
multiple days, the function inserts additional date tags in between these main storm periods if the non-storm periods were sufficiently
long and characterized by substantial accumulations of small daily snowfall amounts. The exact rules and thresholds for establishing
the model-derived date tags are described in detail in the appendix of the paper, or can be viewed directly in the source code below.
}
\examples{

region <- "GNP"
season <- 2022
PDT <- derivePotentialDatetags(WX24)
par(mfrow = c(2, 1))
plot(WX24$date, WX24$hn72, type = "l", col = "red", ylim = c(0, max(WX24$hn24, WX24$hn72, 10.5)))
lines(WX24$date, WX24$hn24)
abline(v = PDT, col = "gray70", lty = "dotted")
abline(h = 10, col = "gray70", lty = "dotted")
legend("topright", c("hn72", "hn24"), col = c("red", "black"), lty = c("solid", "solid"), bty = "n")
mtext(paste(region, season), line = 1)
plot(WX24$date, WX24$rain72, type = "l", col = "red", ylim = c(0, max(WX24$rain72, WX24$rain24, 5.5)))
lines(WX24$date, WX24$rain24)
abline(v = PDT, col = "gray70", lty = "dotted")
abline(h = 5, col = "gray70", lty = "dotted")
legend("topright", c("rain72", "rain24"), col = c("red", "black"), lty = c("solid", "solid"), bty = "n")


}
