% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determinePrevalentGtypes.R
\name{determinePrevalentGtypes}
\alias{determinePrevalentGtypes}
\title{Determine and label the prevalent grain types of simulated unstable layers grouped by model-derived date-tags}
\usage{
determinePrevalentGtypes(layer)
}
\arguments{
\item{layer}{a snowprofileLayers-like object of one or multiple unstable layers from any grid point and any date that are each associated to a datetag.}
}
\value{
This function adds the following columns to the input data.frame:
\itemize{
\item prevalentGtype1Today and prevalentGtype2Today
\item prevalentGtype1 and prevalentGtype2
}
}
\description{
This function takes a data.frame (similar to \code{snowprofileLayers}), but containing layers associated with different
date tags, different grid points, and different points in time. All layers in this data.frame should be unstable
(as determined by your favorite approach to identifying instability, e.g. statistical approach of p_unstable)
and should be associated with a (model-derived) date tag (by \link{assignLayer2Datetag}).
}
\details{
The function determines the prevalent grain types of each group of simulated layers. To be able to label the layer groups
with either a single grain type (FC, SH, or MFcr) or by a mix of two classes (e.g., SH/FC), the function identifies
the two most prevalent grain types within the group of layers associated with a date tag at each grid point for each day.
Then it identifies the two most prevalent grain types across all grid points for each day,
and eventually for the entire lifetime of the layer.

To address the well-known SNOWPACK behaviour of transforming most SH layers into DH layers after several days of burial,
the function judges DH layers as FC layers if only a negligible amount (less than 10\\%) of SH is encountered during the
layer's lifetime. If the fraction of SH layers was higher than 10~\\%, it judges all DH layers as SH layers.

Finally the layer groups are labeled with the resulting one or two prevalent grain types (at each day and over the entire lifetime of the layer).
}
