% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_captWKLquality.R
\name{format_captWKLquality}
\alias{format_captWKLquality}
\title{Format data.frame containing quality/agreement indicators}
\usage{
format_captWKLquality(
  VS,
  mergedVD,
  simulatedGsizeAtBurial = NA,
  WKL_SHonly = NA,
  floorDQ = TRUE
)
}
\arguments{
\item{VS}{the data.frame as returned by evaluate_main.R section 'assessQualityOfCaptWKL'}

\item{mergedVD}{the data.frame as returned by evaluate_main.R section 'mergeVDvframe'}

\item{simulatedGsizeAtBurial}{a data.frame produced in \code{scratch.R}, that contains information about the simulated grain size of layers of concern at their times of burial}

\item{WKL_SHonly}{LEGACY: this is only relevant for fherla's wklvalidation project. If you want to replicate the wklvalidation with new data, this feature is already automated by
having a column in the processingDB table \code{wkl} called \code{SH_only}. This variable will then be used instead of the data.frame provided here, which manually inserts the SH_only flag.
(See \code{scratch.R} for more information about this flag.)}
}
\description{
i.e., apply some filtering, merge with data from mergedVD$wkl, format data classes, compute some interesting variables
}
\seealso{
source code in evaluate_main.R
}
