% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_confusionMatrix.R
\name{format_confusionMatrix}
\alias{format_confusionMatrix}
\title{Format data.frame containing quality/agreement indicators}
\usage{
format_confusionMatrix(CM, mergedVD = NA, WKL_SHonly = NA)
}
\arguments{
\item{CM}{the data.frame as returned by evaluate_main.R section 'deriveConfusionMatrix'}

\item{mergedVD}{the data.frame as returned by evaluate_main.R section 'mergeVDvframe'}

\item{WKL_SHonly}{LEGACY: this is only relevant for fherla's wklvalidation project. If you want to replicate the wklvalidation with new data, this feature is already automated by
having a column in the processingDB table \code{wkl} called \code{SH_only}. This variable will then be used instead of the data.frame provided here, which manually inserts the SH_only flag.
(See \code{scratch.R} for more information about this flag.)}
}
\description{
i.e., apply some filtering, calculating/formatting of gtype_classes, etc.
}
\seealso{
source code in evaluate_main.R
}
