% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_ROCRPerformanceObject.R
\name{format_ROCRPerformanceObject}
\alias{format_ROCRPerformanceObject}
\title{Format ROCR performance object}
\usage{
format_ROCRPerformanceObject(perf, switchXY = FALSE)
}
\description{
This function removes all values that correspond to a threshold criterion of 0 \%, which is
not meaningful in our application
}
\details{
I also tested whether switching x and y values would result in a nicer plot next to the
precision--recall curve with recall always being on the x-axis. (it did not)
}
