% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWLcapturedByModel_subs.R
\name{getMaxElev}
\alias{getMaxElev}
\title{Get max elev of a PWL from ASSESS frame in WklList}
\usage{
getMaxElev(pwl, WklList)
}
\arguments{
\item{pwl}{character name of pwl}

\item{WklList}{~}
}
\value{
numeric scalar
}
\description{
Get max elev of a PWL from ASSESS frame in WklList
}
