% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWLcapturedByModel_subs.R
\name{getRelevantGtype}
\alias{getRelevantGtype}
\title{Extract relevant weak layer grain types for a pwl contained in WklList}
\usage{
getRelevantGtype(pwl, WklList)
}
\arguments{
\item{pwl}{character name of pwl}

\item{WklList}{~}
}
\value{
a list where each element holds a vector of grain types to search (e.g., \code{list[1] = c("SH", "DH")}, \code{list[2] = c("IFsc", "MFcr")}).
The names of the list entries are 'primary', 'secondary', and 'tertiary' (in that order, if not any one is missing).
}
\description{
Due to adjustments of our research strategy when this part of the software had already been developed, it is important to carefully read this description to clearly understand how the analysis deals with gtypes!
In the script(s) \code{scripts/prepareWkl_GNP.R} (and analogous others) each tracked Wkl is associated with a \strong{primary}, \strong{secondary}, and \strong{tertiary} grain type. While the primary \strong{and} secondary grain type
can be thought of as the main grain types of concern (e.g., Jan 17th SH/FC --> primary: SH, secondary: FC; both are eqyally important), the tertiary grain type can be thought of as an additional grain type that might also
be present and potentially also concerning (albeit less) (e.g., Feb 10th SH/MFcr --> primary: SH, tertiary: MFcr). There might be cases of Wkls that are primarily a crust, and tertiarily a SH layer, or vice versa.
}
\details{
Important: The just described nomenclature (applied to the WklList) is changed within this function and therefore in the database output of the validation! Despite which gtype is associated to primary versus tertiary,
this function will \emph{always} make the crust gtype of tertiary rank, so that
\itemize{
\item primary rank refers to all non-crust gtypes mentioned in assessments/ comments (i.e., primary and secondary non-crust gtypes from WklList above)
\item secondary rank refers to all persistent non-crust gtypes (= SH DH FC FCxr) (if not all of these gtypes are already present in the primary rank)
\item tertiary rank refers to all crust gtypes (if present)
}

Example A:
\itemize{
\item \code{WklList[[A]]$ASSESS$GRAIN_TYPE == SH}, \code{WklList[[A]]$ASSESS$GRAIN_TYPE_SECONDARY == NULL}, \code{WklList[[pwl]]$ASSESS$GRAIN_TYPE_TERTIARY == IFsc}
\item --> database vframe.gtype_rank: primary = SH DH, secondary = SH DH FC FCxr, tertiary = IFsc MFcr
}

Example B:
\itemize{
\item \code{WklList[[B]]$ASSESS$GRAIN_TYPE == IFrc}, \code{WklList[[A]]$ASSESS$GRAIN_TYPE_SECONDARY == NULL}, \code{WklList[[pwl]]$ASSESS$GRAIN_TYPE_TERTIARY == FC}
\item --> database vframe.gtype_rank: primary = FC FCxr DH, secondary = FC FCxr DH SH, tertiary = IFrc MFcr
}

Note that all of the ranks can potentially be missing. If a layer is purely a crust layer, primary and secondary will be missing in database. If primary would be identical to secondary, secondary will be missing in the database.
And if there is no crust, tertiary will be missing in the database.

To still know which layer is primarily a crust layer, the database table \code{wkl} contains a column \code{iscrust}, which is set to 1 if \verb{WklList[[X]]$ASSESS$GRAIN_TYPE == MFcr IFrc IFsc}, and 0 otherwise.
}
