% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWLcapturedByModel_subs.R
\name{getSubsetOfSimul}
\alias{getSubsetOfSimul}
\title{Subset relevant snow profiles from a large simulation data set}
\usage{
getSubsetOfSimul(
  simul,
  sm,
  vdate,
  band,
  gtype_class = "NA",
  validateAspects = FALSE,
  max_elev = 9999,
  avprobs = NULL,
  aspects = NULL
)
}
\arguments{
\item{simul}{large snowprofileSet containing entire simulation}

\item{sm}{a summary(simul) containing meta data}

\item{vdate}{validation date as character string "YYYY-MM-DD"}

\item{band}{elevation band character string (ALP, TL, BTL), case insensitive}

\item{gtype_class}{only required if IFsc to constrain aspects}

\item{validateAspects}{extract only flat field profiles (FALSE) or also virtual slopes (TRUE)?}

\item{max_elev}{extract profiles below this elevation}

\item{avprobs}{if validateAspects is TRUE, and you still need to extract the relevant aspects, provide a data frame containing the relevant avalanche problem assessments}

\item{aspects}{if not NULL, validateAspects is automatically set to TRUE. Can provide aspects instead of avprobs.}
}
\value{
a subset snowprofileSet of the input data set
}
\description{
... based on date, band, max_elev, and optionally relevant_gtype and aspects.
}
