% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBinteractions2.R
\name{getVdata}
\alias{getVdata}
\title{Retrieve comprehensive wklvalidation data from DB}
\usage{
getVdata(
  DBCon = NULL,
  DB = "Snow_DB",
  season = NA,
  region = NA,
  band = NA,
  captWKL = TRUE,
  poorWKL = TRUE,
  poorWKLconstraint = "pu >= 0.77 AND gtype NOT IN ('PP', 'DF')",
  studyArea = FALSE,
  limitWKL2vframe = TRUE,
  gtype_rank = c("primary", "secondary", "tertiary"),
  Renviron_path = ifelse(Sys.getenv("USER") == "flo",
    "/home/flo/documents/sfu/code/cedarstuff/CedarTunnel.Renviron",
    "/home/fherla/Cedar.Renviron"),
  verbose = FALSE
)
}
\arguments{
\item{DBCon}{DB connection}

\item{DB}{which DB to connect to? "Snow_DB" / "Steed_DB"}

\item{season}{season(s) you're interested}

\item{region}{region(s) you're interested}

\item{captWKL}{should this table be appended to the output list?}

\item{poorWKL}{should this table be appended to the output list?}

\item{studyArea}{include studyArea table to output list (will be limited to those grid points present in either captWKL or poorWKL)}

\item{limitWKL2vframe}{TRUE/FALSE; False = get all WKL layers from DB}

\item{gtype_rank}{to extract into vframe, captWKL, and poorWKL? see Details below.}

\item{Renviron_path}{path to Renviron file for connection details}
}
\value{
list of tables \code{executions}, \code{vframe}, \code{wkl}, \code{captWKL}, \code{poorWKL}
}
\description{
This function retrieves the vframe joined with execution table, and corresponding captWKLs & poorWKLs
}
\details{
remote usage prerequisite: on a terminal keep running SSH tunnel:
\verb{ssh -L 8888:cedar-mysql-vm.int.cedar.computecanada.ca:3306 fherla@cedar.computecanada.ca}

For gtype_rank choices: You can retrieve all crust entries from the DB by specifying \code{gtype_rank = 'tertiary'}. If you also specify \code{'primary'}, you will get all persistent grain types that were listed in the
forecasters' assessments. If you specify \code{'secondary'} instead, you will get all persistent grain types whether they were mentioned in the assessments or not (i.e., a relaxed prerequisite for the validation).
Note, however, that 'secondary' ranks only exist where 'primary' ranks not already contain all persistent grain types. Therefore you can specify both \code{c('primary', 'secondary')} to retrieve secondary ranks when
available and primary ones otherwise. If you set gtype_rank to NULL, you will get all DB entries despite the gtype_rank. Note that this is only possible when retrieving only the vframe without captWKLs and poorWKLs.
}
