% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBinteractions.R
\name{getWKLuuid}
\alias{getWKLuuid}
\title{Retrieve wkl_uuids from DB}
\usage{
getWKLuuid(DBCon, wkl_name, region = NULL)
}
\arguments{
\item{DBCon}{DB connection}

\item{wkl_name}{wkl_name of format 'season.name'}

\item{region}{DB query condition}
}
\value{
data.frame
}
\description{
This function retrieves the wkl_uuids alongside the wkl_name, season, region, dq from the \code{wkl} table.
}
