% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWLcapturedByModel_subs.R
\name{getWxTimeWindow}
\alias{getWxTimeWindow}
\title{Calculate PWL formation time window around burial date based on wx data}
\usage{
getWxTimeWindow(
  profile,
  pwl_bdate,
  wx_daily,
  timewindow_lower = -30,
  timewindow_upper = 3
)
}
\arguments{
\item{profile}{only required to obtain station_id}

\item{pwl_bdate}{burial date of PWL}

\item{wx_daily}{data frame of daily resampled wx data (need \verb{$rain24}, \verb{$hn24}, \verb{$station_id})}

\item{timewindow_lower}{max window size before human burial date tag}

\item{timewindow_upper}{max window size after human burial date tag}
}
\value{
vector of length 2, providing lower and upper time window limits, respectively.
}
\description{
Max time window as provided by arguments will be made tighter if cumulative hn24 > 10 cm or cumulative rain24 > 5 mm.
}
