% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctree_plottingSupport.R
\name{innerWeights}
\alias{innerWeights}
\title{Define color, shapes, and labels for inner ctree nodes}
\usage{
innerWeights(
  obj,
  id = TRUE,
  pval = TRUE,
  abbreviate = FALSE,
  fill = "white",
  gp = gpar()
)
}
\description{
Define color, shapes, and labels for inner ctree nodes
}
