% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBinteractions.R
\name{insertTable}
\alias{insertTable}
\title{Custom insertTable command}
\usage{
insertTable(DBCon, DF, tblname)
}
\arguments{
\item{DBCon}{DB connection}

\item{DF}{data.frame in correct order of columns, correct colnames, and correct data types (see Description)}

\item{tblname}{table to insert into (must exist)}
}
\value{
Returns TRUE if successful, returns the error silently otherwise. The query that caused the error is stored in \verb{$query}!
Tip: Wrap function call in tryCatch to catch error and informative traceback etc..
}
\description{
Equivalent to \code{RMySQL::dbWriteTable}, appends into existing table, but issues error if query can't fully be executed.
Note that the data.frame to be inserted needs to be named correctly and columns need to be ordered correctly. Data types correctly.
Character arrays will be quoted automatically, NA/NaN/NULL values will be handled automatically, and factors will be converted to characters.
}
