% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeVDvframe.R
\name{mergeVDvframe}
\alias{mergeVDvframe}
\title{Merge two database tables \code{vframe}}
\usage{
mergeVDvframe(
  VD,
  VD2add,
  add_gsizeAtBurial = FALSE,
  add_timewindows = FALSE,
  add_poorWKLgtypeclass = TRUE,
  region = NA,
  band = NA,
  gtype_rank = NA
)
}
\arguments{
\item{VD}{ValidationData object as obtained from \link{getVdata}}

\item{VD2add}{ValidationData object to be merged into VD}

\item{add_gsizeAtBurial}{logical switch to compute median simulated grain size at time of burial
(and include it in vframe) with a function call to \link{computeGsizeAtBurial}}

\item{add_timewindows}{logical switch to compute median length of dry period before burial
(and include it in vframe) with a function call to \link{computeLengthOfPreceedingDrySpell}}

\item{add_poorWKLgtypeclass}{logical switch to compute and label the grain type classes of all unstable layers
that are not associated with a layer of human concern (i.e., table \code{poorWKL}). This is carried out with a
call to \link{determindePrevalentGtypes} and then assembling the returned gtype labels in a format consistent
with how gtypes have been handled in the paper.}

\item{region}{info from other tables (required only if add_poorWKLgtypeclass = TRUE)}

\item{band}{info from other tables (required only if add_poorWKLgtypeclass = TRUE)}

\item{gtype_rank}{info from other tables (required only if add_poorWKLgtypeclass = TRUE)}
}
\value{
object VD (tables \code{executions}, \code{vframe}, \code{wkl}) merged with the corresponding tables from VD2add
}
\description{
The ValidationData \code{VD} contains several different tables from the \code{wklvalidation} database. \code{vframe} is one of them,
containing a summary of captured layers at a given day for differet layers of concern. For the final analysis
it is necessary to have the data stored in a long vframe table that contains all the rows from all layers of concern,
all elevation bands, regions, dates, etc. However, for each layer of concern several preprocessing steps need
to be carried out when reading the data from the database (this is all done in \code{evaluate_main.R}). This function
facilitates merging all the individual vframe tables back together! In fact, it also merges the tables \code{executions}
and \code{wkl}, but discards the tables with individual layer information (which are simply too big!) \code{captWKL} and \code{poorWKL}.
}
\details{
Given the workflow in \code{evaluate_main.R}, information from individual layers is available up to this function,
but not afterwards. So, this function includes several features to compute summary statistics from individual
layers (e.g., simulated grain size at burial, etc).
}
