% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_mvar.R
\name{node_mvar_violin}
\alias{node_mvar_violin}
\title{Plot terminal nodes of multivariate ctrees as violins}
\usage{
node_mvar_violin(
  obj,
  which = NULL,
  id = TRUE,
  pop = TRUE,
  ylines = NULL,
  mainlab = NULL,
  varlab = TRUE,
  bg = "white",
  terminal_panel_mvar = node_violinplot,
  ...
)
}
\description{
Can be used exactly as the partykit function \code{node_mvar}.
}
\seealso{
\link{node_violinplot}
}
\author{
fherla
}
