% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_violinplot.R
\name{node_violinplot}
\alias{node_violinplot}
\title{Plot terminal nodes of ctrees as violins}
\usage{
node_violinplot(
  obj,
  col = "black",
  fill = "lightgray",
  bg = "white",
  width = 0.8,
  yscale = NULL,
  ylines = 3,
  cex = 0.5,
  id = TRUE,
  mainlab = NULL,
  gp = gpar(),
  col.box = "black",
  fill.box = "black",
  fill.median = "white"
)
}
\description{
Can be used exactly as other partykit functions for terminal nodes, e.g. \code{node_boxplot}.
}
\seealso{
\link{node_mvar_violin}
}
\author{
fherla
}
