% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessValidation.R
\name{plotSeasonalOverviewProblemsPoorWKL}
\alias{plotSeasonalOverviewProblemsPoorWKL}
\title{Timeseries seasonal overview avy problems and poor WKLs}
\usage{
plotSeasonalOverviewProblemsPoorWKL(
  VD,
  showPWLs = FALSE,
  shadeAvProbPeriods = TRUE,
  plotCircles = FALSE,
  plotBars = !plotCircles,
  pwcex = 1.5,
  withXAxis = TRUE,
  xticks_Sat = TRUE,
  captCol = RColorBrewer::brewer.pal(5, "Pastel1"),
  cex.tag = par()$cex.axis,
  ...
)
}
\description{
Timeseries seasonal overview avy problems and poor WKLs
}
