% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessValidation.R
\name{plotTSaggregatedWKL}
\alias{plotTSaggregatedWKL}
\title{Time series of summarized/aggregated WKLs}
\usage{
plotTSaggregatedWKL(vdat, wkl_uuid = NA, gclasses = "all", xage = "all", ...)
}
\description{
Time series of summarized/aggregated WKLs
}
