% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTSAvProbLikelihood.R
\name{plotTSAvProbLikelihood}
\alias{plotTSAvProbLikelihood}
\title{Plot avalanche problem likelihood of specific avalanche problem types}
\usage{
plotTSAvProbLikelihood(
  Bull,
  ElevBand = "Tl",
  ProbOrder = c("Persistent Slabs", "Deep Persistent Slabs", "Storm Slabs"),
  withXAxis = TRUE,
  withYAxis = TRUE,
  xticks_Sat = TRUE,
  plotBars = FALSE,
  ...
)
}
\arguments{
\item{Bull}{Bulletin object of the desired region/season}

\item{ElevBand}{elevation band of interest (must be contained in \code{Bull})}

\item{ProbOrder}{Which problem types to plot and if there are multiple at any given day, which one should be plotted with higher priority?}

\item{withXAxis}{plot x axis}

\item{withYAxis}{plot y axis}

\item{xticks_Sat}{xticks on each Saturday?}

\item{plotBars}{plot barchart? if not, points and recatangles are plotted for typ/min/max values.}
}
\description{
This function is used to plot the avalanche problem likelihood of storm and persistent problems
in a barchart-way that looks similar to the barchart of proportion_unstable plots (e.g., Fig. 14c in paper).
}
\details{
Instead of plotting only these three problem types, the function can be customized to plot any problem type of interest.
Since it plots one time series, it can only plot one problem type per day. If there are more problems issued at that
day, the argument \code{ProbOrder} defines which one is plotted.
}
