% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessValidation.R
\name{plotTSWKL}
\alias{plotTSWKL}
\title{Time series of WKL}
\usage{
plotTSWKL(
  vdat,
  wkl_uuid,
  plotIndices = ifelse(length(wkl_uuid) == 1, TRUE, FALSE),
  cex = 1
)
}
\description{
Time series of WKL
}
