% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWLcapturedByModel_subs.R
\name{PWLcapturedByIndividualProfile}
\alias{PWLcapturedByIndividualProfile}
\title{Confirm whether or not a specific layer exists in an individual profile}
\usage{
PWLcapturedByIndividualProfile(x, gtypes, pwl_bdate)
}
\arguments{
\item{x}{snowprofile with additional fields timelimit_lower and timelimit_upper that specify the beginning of ddate time window (which goes up until pwl_bdate)
and the end of bdate time window (which starts 1 day before pwl_bdate)}

\item{gtypes}{to search for}

\item{pwl_bdate}{from WklList}
}
\value{
indices of captured layers
}
\description{
Confirm whether or not a specific layer exists in an individual profile
}
