% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWLcapturedByModel.R
\name{PWLcapturedByModel}
\alias{PWLcapturedByModel}
\title{Check whether persistent weak layer is captured in model simulations}
\usage{
PWLcapturedByModel(
  WklList,
  simul,
  wx_daily = NA,
  sm = summary(simul),
  bulletins = NA,
  when = "day 3",
  validateAspects = FALSE,
  timewindow_lower = -30,
  timewindow_upper = 3,
  savePNG_prefix = NA,
  verbose = TRUE,
  DBCon = NULL,
  region = NA,
  exec_uuid = ifelse(is.null(DBCon), NA,
    stop("Must provide exec_uuid if writing to DB!")),
  saveCapturedWKL = ifelse(is.null(DBCon), FALSE, TRUE),
  saveVframe = ifelse(is.null(DBCon), FALSE, TRUE)
)
}
\arguments{
\item{WklList}{a list of PWLs as extracted by \code{SarpBulletinTools::getWklInfoForBulletins}. Note that this routine will put \verb{$GRAIN_TYPE} and \verb{$GRAIN_TYPE_SECONDARY} into one
gtype class and \verb{$GRAIN_TYPE_TERTIARY} into another one. See code for further clarification.}

\item{simul}{a \code{snowprofileSet} containing the simulations. The function will assess every elevation band contained in \code{simul}.}

\item{wx_daily}{a data.frame containing weather information which are aggregated to 24h intervals (required fields are \verb{$hn24} and \verb{$rain24}). If \code{wx_daily} is provided, the timewindow to
search for PWLs in the simulation is modified based on the underlying weather conditions. That means, \code{timewindow_lower} and \code{timewindow_upper} are pruned if hn24 (rain24) exceeds 10 cm (5 mm).
This feature is ignored when \code{wx_daily} is \code{NA}.}

\item{sm}{a summary data.frame for simul}

\item{bulletins}{Bulletins object that will be used to extract danger rating. Must only contain one danger rating per day and elevation band (i.e., only contain one region!) (optional)}

\item{when}{timing mode of validation. Choose between \code{"day #"}, \code{"first problem"}, \code{"first persistent problem"}, \code{"first deep persistent problem"}, or \code{"each problem day"}}

\item{validateAspects}{boolean switch to consider aspects while validating (i.e., extract hazardous aspects and search in corresponding profiles)}

\item{timewindow_lower}{scalar of unit 'days' before the reported weak layer datetag to start ddate search window in simulation}

\item{timewindow_upper}{scalar of unit 'days' after the reported weak layer datetag to end bdate search window in simulation}

\item{savePNG_prefix}{prefix to manual validation verification figure, such as directory and prefix to filename. automatic filename postfix will include pwlidx, bdate, vdate, & gtype_class. e.g.,
\code{"/home/flo/documents/sfu/code/rpackages_SARP/sarp.2021.herla.snowprofilevalidation/output/figures/validationVerificationPlots/gnp/"}. If \code{NA} no figure will be printed.}

\item{verbose}{print messages related to function progress?}

\item{DBCon}{an established DB connection. If NULL, no results will be written to a database (i.e., dry run)}

\item{region}{if DBCon provided, region must be supplied, too. e.g., "GLACIER"}

\item{exec_uuid}{execution UUID for DB management}

\item{saveCapturedWKL}{write captured layers to DB?}

\item{saveVframe}{write validation data.frame to DB?}
}
\value{
A data frame that contains the validation information. Additionally, the routine can plot manual validation verification figures to manually check whether the weak layers
were correctly deemed captured/not captured. And finally, the routine can save all captured layers to a mysql database either at Cedar or at steed (local fherla).
}
\description{
This function validates whether PWLs reported in avalanche forecasting assessments also exist in model simulations.
In the paper flowchart (Fig. 2a), this function corresponds to the matching step at the grid point level.
}
