% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWLencounteredInModel.R
\name{PWLencounteredInModel}
\alias{PWLencounteredInModel}
\title{Finds weak layers in simulation and stores them in DB}
\usage{
PWLencounteredInModel(
  simul,
  sm = summary(simul),
  DBCon = NULL,
  savePoorWKL = ifelse(is.null(DBCon), FALSE, TRUE),
  exec_uuid = ifelse(is.null(DBCon), NA,
    stop("Must provide exec_uuid if writing to DB!")),
  verbose = TRUE
)
}
\arguments{
\item{simul}{snowprofileSet}

\item{sm}{simulation meta data summary}

\item{DBCon}{DB connection}

\item{savePoorWKL}{boolean switch to write to \code{fherla_wklvalidation.poorWKL}}

\item{exec_uuid}{execution uuid to reference calculation}
}
\value{
writes to DB
}
\description{
In the paper flowchart (Fig. 2a), this function does not correspond to a specific step, but instead it finds all 'Unstable Layers' in the simulations.
}
