% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegionNamesList.R
\docType{data}
\name{RegionNamesList}
\alias{RegionNamesList}
\title{List with studyArea names in different formats}
\format{
An object of class \code{list} of length 3.
}
\usage{
RegionNamesList
}
\description{
List with studyArea names in different formats
}
\keyword{datasets}
