% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBinteractions.R
\name{saveCapturedLayers2DB}
\alias{saveCapturedLayers2DB}
\title{Save all captured layers to mysql database at cedar/steed}
\usage{
saveCapturedLayers2DB(profile, idx, DBCon, extdata)
}
\arguments{
\item{profile}{profile which contains captured layers}

\item{idx}{indices of captured layers which will be saved to DB}

\item{DBCon}{DB connection}

\item{extdata}{1 row data.frame with columns exec_uuid, wkl_uuid, vf_uuid, vdate}
}
\value{
no return, but writes data to DB
}
\description{
Save all captured layers to mysql database at cedar/steed
}
