% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBinteractions.R
\name{saveExecutionsTable2DB}
\alias{saveExecutionsTable2DB}
\title{Write execution information to mysql database at cedar/steed}
\usage{
saveExecutionsTable2DB(DBCon, execFrame)
}
\arguments{
\item{DBCon}{DB connection}

\item{execFrame}{data frame 1 row including execution details}
}
\value{
no return, but writes to DB
}
\description{
Write execution information to mysql database at cedar/steed
}
