% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBinteractions.R
\name{savePoorLayers2DB}
\alias{savePoorLayers2DB}
\title{Save all layers with poor stability to mysql database at cedar/steed}
\usage{
savePoorLayers2DB(lyrs, idx, DBCon, extdata)
}
\arguments{
\item{lyrs}{(stacked) profile layers incl. station(_id), angle, aspect}

\item{idx}{indices of poor layers which will be saved to DB}

\item{DBCon}{DB connection}

\item{extdata}{1 row data.frame with columns exec_uuid, vdate, nSP}
}
\value{
no return, but writes data to DB
}
\description{
Save all layers with poor stability to mysql database at cedar/steed
}
