% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTSAvProbLikelihood.R
\name{shadeAvProbPeriod}
\alias{shadeAvProbPeriod}
\title{Shade period with specific avalanche problem types in their problem color}
\usage{
shadeAvProbPeriod(
  Bull,
  ElevBand = "Tl",
  ProbOrder = c("Persistent Slabs", "Deep Persistent Slabs", "Storm Slabs")
)
}
\description{
This function is the equivalent to \link{plotTSAvProbLikelihood}, shading periods with specific problems in their native color.
}
