% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transactPoor2CaptWKL.R
\name{transactPoor2CaptWKL}
\alias{transactPoor2CaptWKL}
\title{Regional check of layer matching at grouped level}
\usage{
transactPoor2CaptWKL(VD, WX24, potentialDatetags = NULL)
}
\arguments{
\item{VD}{ValidationData object as obtained from \link{getVdata}}

\item{WX24}{a data.frame with weather data containing only one record per day. Each record must include hn24 (75th percentile), hn72 (25th percentile),
rain24 (75th percentile) and rain72 (25th percentile).}

\item{potentialDatetags}{as obtained from \link{derivePotentialDatetags}. Not required if WX24 provided.}
}
\value{
the VD object with slightly changed table contents and recomputed summary statistics of the relevant variables.
}
\description{
In the paper flowchart (Fig. 2a), this function corresponds to the matching step at the regionally grouped level. It is implemented in a way
that first the grid point level matching is carried out, and the resulting layers are written to two DB tables \code{captWKL} and \code{poorWKL}. Then
this function is intended to perform the regional check after reading a data subset from the database. All layers that are found to be associated
with the seemingly wrong data table are moved between the tables (i.e., some poorWKL layers are moved over to the captWKL table) and relevant
summary statistics stored in other tables (such as vframe) are recomputed.
}
\details{
It is implemented in this way, because this regional check was designed many months after the grid point level matching was already in place.
}
