% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWLcapturedByModel_subs.R
\name{vizCapturedPWL}
\alias{vizCapturedPWL}
\title{Vizualize and manually verify whether layer is captured or not}
\usage{
vizCapturedPWL(
  WklList,
  simul = NA,
  meta = NA,
  validateAspects = FALSE,
  vf = NA,
  pwl = vf$pwl,
  vdate = vf$validation_date,
  band = vf$band,
  gtype_class = vf$gtype_class,
  timelimit_lower = vf$timewindow_lower,
  timelimit_upper = vf$timewindow_upper,
  relevant_gtypes = getRelevantGtype(pwl, WklList)[[gtype_class]],
  gtype_rank = NA,
  simul_subset = NA,
  wx_daily = NA,
  pwl_exists = NA,
  savePNG = FALSE,
  savePNG_prefix = NA,
  verbose = TRUE,
  keep.results = TRUE
)
}
\description{
This function can also be used to reconstruct the simulation subset at a later point
}
